# check_manager/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Check Manager API",
    "version": "v1"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIOKResponseList_NoneType_"
                }
              }
            }
          }
        }
      }
    },
    "/v1/check_templates/": {
      "get": {
        "summary": "Get Check Templates",
        "operationId": "get_check_templates_v1_check_templates__get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "restrict IDs to include",
              "title": "Ids"
            },
            "description": "restrict IDs to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIOKResponseList_CheckTemplateAttributes_"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/check_templates/{check_template_id}": {
      "get": {
        "summary": "Get Check Template",
        "operationId": "get_check_template_v1_check_templates__check_template_id__get",
        "parameters": [
          {
            "name": "check_template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIOKResponse_CheckTemplateAttributes_"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/checks/": {
      "get": {
        "summary": "Get Checks",
        "operationId": "get_checks_v1_checks__get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "restrict IDs to include",
              "title": "Ids"
            },
            "description": "restrict IDs to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIOKResponseList_OutCheckAttributes_"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      },
      "post": {
        "summary": "Create Check",
        "operationId": "create_check_v1_checks__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/InCheck"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIOKResponse_OutCheckAttributes_"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/checks/{check_id}": {
      "get": {
        "summary": "Get Check",
        "operationId": "get_check_v1_checks__check_id__get",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIOKResponse_OutCheckAttributes_"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      },
      "delete": {
        "summary": "Remove Check",
        "operationId": "remove_check_v1_checks__check_id__delete",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "APIErrorResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "errors"
        ],
        "title": "APIErrorResponse"
      },
      "APIOKResponseList_CheckTemplateAttributes_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Resource_CheckTemplateAttributes_"
            },
            "type": "array",
            "title": "Data"
          },
          "links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIOKResponseList[CheckTemplateAttributes]"
      },
      "APIOKResponseList_NoneType_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Resource_NoneType_"
            },
            "type": "array",
            "title": "Data"
          },
          "links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIOKResponseList[NoneType]"
      },
      "APIOKResponseList_OutCheckAttributes_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Resource_OutCheckAttributes_"
            },
            "type": "array",
            "title": "Data"
          },
          "links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIOKResponseList[OutCheckAttributes]"
      },
      "APIOKResponse_CheckTemplateAttributes_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Resource_CheckTemplateAttributes_"
          },
          "links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIOKResponse[CheckTemplateAttributes]"
      },
      "APIOKResponse_OutCheckAttributes_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Resource_OutCheckAttributes_"
          },
          "links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIOKResponse[OutCheckAttributes]"
      },
      "CheckTemplateAttributes": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/Json"
          },
          "arguments": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "object"
              }
            ],
            "title": "Arguments"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "arguments"
        ],
        "title": "CheckTemplateAttributes"
      },
      "Error": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "status",
          "code",
          "title"
        ],
        "title": "Error"
      },
      "ErrorSource": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ErrorSourcePointer"
          },
          {
            "$ref": "#/components/schemas/ErrorSourceParameter"
          },
          {
            "$ref": "#/components/schemas/ErrorSourceHeader"
          }
        ]
      },
      "ErrorSourceHeader": {
        "properties": {
          "header": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Header"
          }
        },
        "type": "object",
        "required": [
          "header"
        ],
        "title": "ErrorSourceHeader"
      },
      "ErrorSourceParameter": {
        "properties": {
          "parameter": {
            "type": "string",
            "title": "Parameter"
          }
        },
        "type": "object",
        "required": [
          "parameter"
        ],
        "title": "ErrorSourceParameter"
      },
      "ErrorSourcePointer": {
        "properties": {
          "pointer": {
            "type": "string",
            "title": "Pointer"
          }
        },
        "type": "object",
        "required": [
          "pointer"
        ],
        "title": "ErrorSourcePointer"
      },
      "InCheck": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InCheckData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "InCheck"
      },
      "InCheckAttributes": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/InCheckMetadata"
          },
          "schedule": {
            "type": "string",
            "title": "Schedule"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "schedule"
        ],
        "title": "InCheckAttributes"
      },
      "InCheckData": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "check"
          },
          "attributes": {
            "$ref": "#/components/schemas/InCheckAttributes"
          }
        },
        "type": "object",
        "required": [
          "attributes"
        ],
        "title": "InCheckData"
      },
      "InCheckMetadata": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "template_args": {
            "$ref": "#/components/schemas/Json"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "template_id",
          "template_args"
        ],
        "title": "InCheckMetadata"
      },
      "Json": {
        "type": "object"
      },
      "Link": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "$ref": "#/components/schemas/LinkObject"
          }
        ]
      },
      "LinkObject": {
        "properties": {
          "href": {
            "type": "string",
            "title": "Href"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "href"
        ],
        "title": "LinkObject"
      },
      "Links": {
        "properties": {
          "self": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Link"
              },
              {
                "type": "null"
              }
            ]
          },
          "describedby": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Link"
              },
              {
                "type": "null"
              }
            ]
          },
          "root": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Link"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "Links"
      },
      "OutCheckAttributes": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/OutCheckMetadata"
          },
          "schedule": {
            "type": "string",
            "title": "Schedule"
          },
          "outcome_filter": {
            "$ref": "#/components/schemas/Json"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "schedule",
          "outcome_filter"
        ],
        "title": "OutCheckAttributes"
      },
      "OutCheckMetadata": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "template_args": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Json"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "OutCheckMetadata"
      },
      "Resource_CheckTemplateAttributes_": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "attributes": {
            "$ref": "#/components/schemas/CheckTemplateAttributes"
          },
          "links": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/Link"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "attributes"
        ],
        "title": "Resource[CheckTemplateAttributes]"
      },
      "Resource_NoneType_": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "attributes": {
            "type": "null",
            "title": "Attributes"
          },
          "links": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/Link"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "attributes"
        ],
        "title": "Resource[NoneType]"
      },
      "Resource_OutCheckAttributes_": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "attributes": {
            "$ref": "#/components/schemas/OutCheckAttributes"
          },
          "links": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/Link"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "attributes"
        ],
        "title": "Resource[OutCheckAttributes]"
      }
    }
  }
}